chore: separate `rust-toolkit`
