+++ packages/rust-toolkit/source/commands.ts
@@ -1,0 +1,4 @@
+/** Languages */
+
+// Rust
+export const RUST_SET_TARGET = 'rust.setTarget';
