+++ packages/rust-toolkit/source/crates/core/Dependency.ts
@@ -5,10 +5,10 @@ import Item from "./Item";
  * Dependency is a data structure to define parsed dependency index, versions and error
  */
 export default interface Dependency {
-  item: Item;
-  versions?: Array<string>;
-  error?: string;
+	item: Item;
+	versions?: Array<string>;
+	error?: string;
 
-  versionCompletionItems?: CompletionList;
-  featureCompletionItems?: Map<string, CompletionList>;
+	versionCompletionItems?: CompletionList;
+	featureCompletionItems?: Map<string, CompletionList>;
 }
