+++ packages/rust-toolkit/tsconfig.json
@@ -1,0 +1,6 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "./build"
+	}
+}
