+++ packages/vscode/extensions/commands.ts
@@ -1,8 +1,3 @@
-/** Languages */
-
-// Rust
-export const RUST_SET_TARGET = 'foretag-rust.setRustTarget';
-
 /** Security */
 
 // Env
